Reorganise store tests into submodules
