+++ p2poolv2_lib/src/store/user.rs
@@ -144,3 +144,144 @@ impl Store {
         Ok(results)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use tempfile::tempdir;
+
+    #[test]
+    fn test_store_and_get_user() {
+        let temp_dir = tempdir().unwrap();
+        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
+
+        let btcaddress = "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa".to_string();
+
+        // Store a new user
+        let user_id = store.add_user(btcaddress.clone()).unwrap();
+
+        // Get user by ID
+        let stored_user = store.get_user_by_id(user_id).unwrap().unwrap();
+        assert_eq!(stored_user.user_id, user_id);
+        assert_eq!(stored_user.btcaddress, btcaddress);
+        assert!(stored_user.created_at > 0);
+
+        // Get user by btcaddress
+        let user_by_address = store.get_user_by_btcaddress(&btcaddress).unwrap().unwrap();
+        assert_eq!(user_by_address.user_id, user_id);
+        assert_eq!(user_by_address.btcaddress, btcaddress);
+
+        // Store same user again - should return same ID
+        let same_user_id = store.add_user(btcaddress.clone()).unwrap();
+        assert_eq!(same_user_id, user_id);
+
+        // Store different user - should get new ID
+        let btcaddress2 = "1BvBMSEYstWetqTFn5Au4m4GFg7xJaNVN2".to_string();
+        let _user_id2 = store.add_user(btcaddress2.clone()).unwrap();
+
+        // Verify both users exist
+        let user1 = store.get_user_by_btcaddress(&btcaddress).unwrap().unwrap();
+        let user2 = store.get_user_by_btcaddress(&btcaddress2).unwrap().unwrap();
+        assert_ne!(user1.user_id, user2.user_id);
+    }
+
+    #[test]
+    fn test_get_nonexistent_user() {
+        let temp_dir = tempdir().unwrap();
+        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
+
+        // Test getting non-existent user by ID
+        let user = store.get_user_by_id(999).unwrap();
+        assert!(user.is_none());
+
+        // Test getting non-existent user by btcaddress
+        let user = store.get_user_by_btcaddress("nonexistent_address").unwrap();
+        assert!(user.is_none());
+    }
+
+    #[test]
+    fn test_user_serialization() {
+        let temp_dir = tempdir().unwrap();
+        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
+
+        let btcaddress = "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa".to_string();
+
+        // Store user
+        let user_id = store.add_user(btcaddress.clone()).unwrap();
+
+        // Retrieve and verify data integrity
+        let stored_user = store.get_user_by_id(user_id).unwrap().unwrap();
+
+        // Verify all fields are correctly serialized/deserialized
+        assert_eq!(stored_user.user_id, user_id);
+        assert_eq!(stored_user.btcaddress, btcaddress);
+        assert!(stored_user.created_at > 0);
+
+        // Verify timestamps are reasonable (within last minute)
+        let now = std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .unwrap()
+            .as_secs();
+        assert!(stored_user.created_at <= now);
+        assert!(stored_user.created_at > now - 60);
+    }
+
+    #[test]
+    fn test_get_btcaddresses_for_user_ids() {
+        let temp_dir = tempdir().unwrap();
+        let store = Store::new(temp_dir.path().to_str().unwrap().to_string(), false).unwrap();
+
+        // Store multiple users
+        let btcaddress1 = "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa".to_string();
+        let btcaddress2 = "1BvBMSEYstWetqTFn5Au4m4GFg7xJaNVN2".to_string();
+        let btcaddress3 = "1QGTJkBFhCjPHqbnwK6z7JfEHefq6Yj2jJ".to_string();
+
+        let user_id1 = store.add_user(btcaddress1.clone()).unwrap();
+        let user_id2 = store.add_user(btcaddress2.clone()).unwrap();
+        let user_id3 = store.add_user(btcaddress3.clone()).unwrap();
+
+        // Test getting btcaddresses for existing user IDs
+        let user_ids = &[user_id1, user_id2, user_id3];
+        let results = store.get_btcaddresses_for_user_ids(user_ids).unwrap();
+
+        assert_eq!(results.len(), 3);
+
+        // Convert to HashMap for easier lookup
+        let result_map: std::collections::HashMap<u64, String> = results.into_iter().collect();
+
+        assert_eq!(result_map.get(&user_id1), Some(&btcaddress1));
+        assert_eq!(result_map.get(&user_id2), Some(&btcaddress2));
+        assert_eq!(result_map.get(&user_id3), Some(&btcaddress3));
+
+        // Test with subset of user IDs
+        let subset_ids = &[user_id1, user_id3];
+        let subset_results = store.get_btcaddresses_for_user_ids(subset_ids).unwrap();
+
+        assert_eq!(subset_results.len(), 2);
+        let subset_map: std::collections::HashMap<u64, String> =
+            subset_results.into_iter().collect();
+
+        assert_eq!(subset_map.get(&user_id1), Some(&btcaddress1));
+        assert_eq!(subset_map.get(&user_id3), Some(&btcaddress3));
+        assert!(!subset_map.contains_key(&user_id2));
+
+        // Test with non-existent user IDs
+        let nonexistent_ids = &[9999, 8888];
+        let empty_results = store
+            .get_btcaddresses_for_user_ids(nonexistent_ids)
+            .unwrap();
+
+        assert_eq!(empty_results.len(), 0);
+
+        // Test with mixed existing and non-existent IDs
+        let mixed_ids = &[user_id1, 9999, user_id2];
+        let mixed_results = store.get_btcaddresses_for_user_ids(mixed_ids).unwrap();
+
+        assert_eq!(mixed_results.len(), 2);
+        let mixed_map: std::collections::HashMap<u64, String> = mixed_results.into_iter().collect();
+
+        assert_eq!(mixed_map.get(&user_id1), Some(&btcaddress1));
+        assert_eq!(mixed_map.get(&user_id2), Some(&btcaddress2));
+        assert!(!mixed_map.contains_key(&9999));
+    }
+}
